+++ backend/authentication/providers/basic/basic.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"net/url"
 
-	"github.com/sensu/sensu-go/api/core/v2"
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/authentication/jwt"
-	"github.com/sensu/sensu-go/backend/authentication/providers"
 	"github.com/sensu/sensu-go/backend/store"
 )
 
@@ -17,10 +17,13 @@ const Type = "basic"
 // Provider represents the basic internal authentication provider
 type Provider struct {
 	Store store.Store
+
+	// ObjectMeta contains the name, namespace, labels and annotations
+	corev2.ObjectMeta `json:"metadata"`
 }
 
 // Authenticate a user, with the provided credentials, against the Sensu store
-func (p *Provider) Authenticate(ctx context.Context, username, password string) (*v2.Claims, error) {
+func (p *Provider) Authenticate(ctx context.Context, username, password string) (*corev2.Claims, error) {
 	if username == "" || password == "" {
 		return nil, errors.New("the username and the password must not be empty")
 	}
@@ -42,7 +45,7 @@ func (p *Provider) Authenticate(ctx context.Context, username, password string) 
 }
 
 // Refresh the claims of a user
-func (p *Provider) Refresh(ctx context.Context, providerClaims v2.ProviderClaims) (*v2.Claims, error) {
+func (p *Provider) Refresh(ctx context.Context, providerClaims corev2.AuthProviderClaims) (*corev2.Claims, error) {
 	user, err := p.Store.GetUser(ctx, providerClaims.UserID)
 	if err != nil {
 		return nil, err
@@ -62,9 +65,14 @@ func (p *Provider) Refresh(ctx context.Context, providerClaims v2.ProviderClaims
 	return claims, nil
 }
 
-// GetName returns the provider name
-func (p *Provider) GetName() string {
-	return "default"
+// GetObjectMeta returns the provider metadata
+func (p *Provider) GetObjectMeta() corev2.ObjectMeta {
+	return p.ObjectMeta
+}
+
+// Name returns the provider name
+func (p *Provider) Name() string {
+	return p.ObjectMeta.Name
 }
 
 // Type returns the provider type
@@ -72,9 +80,23 @@ func (p *Provider) Type() string {
 	return Type
 }
 
-func (p *Provider) claims(username string) v2.ProviderClaims {
-	return v2.ProviderClaims{
-		ProviderID: providers.ID(p),
+// URIPath returns the path component of the basic provider
+func (p *Provider) URIPath() string {
+	return fmt.Sprintf("/api/core/v2/providers/%s/%s",
+		url.PathEscape(Type),
+		url.PathEscape(p.Name()),
+	)
+}
+
+// Validate validates the basic provider configuration
+func (p *Provider) Validate() error {
+	//TODO(palourde): Implement this!
+	return nil
+}
+
+func (p *Provider) claims(username string) corev2.AuthProviderClaims {
+	return corev2.AuthProviderClaims{
+		ProviderID: corev2.AuthProviderID(p),
 		UserID:     username,
 	}
 }