+++ backend/store/store.go
@@ -12,18 +12,6 @@ import (
 	"github.com/sensu/sensu-go/types"
 )
 
-const (
-	// WatchUnknown indicates that we received an unknown watch even tytpe
-	// from etcd.
-	WatchUnknown WatchActionType = iota
-	// WatchCreate indicates that an object was created.
-	WatchCreate
-	// WatchUpdate indicates that an object was updated.
-	WatchUpdate
-	// WatchDelete indicates that an object was deleted.
-	WatchDelete
-)
-
 // ErrAlreadyExists is returned when an object already exists
 type ErrAlreadyExists struct {
 	Key string
@@ -92,36 +80,11 @@ func (e *ErrInternal) Error() string {
 	return fmt.Sprintf("internal error: %s", e.Message)
 }
 
-// WatchActionType indicates what type of change was made to an object in the store.
-type WatchActionType int
-
-func (t WatchActionType) String() string {
-	var s string
-	switch t {
-	case WatchUnknown:
-		s = "Unknown"
-	case WatchCreate:
-		s = "Create"
-	case WatchDelete:
-		s = "Delete"
-	case WatchUpdate:
-		s = "Update"
-	}
-	return s
-}
-
 // A WatchEventCheckConfig contains the modified store object and the action that occured
 // during the modification.
 type WatchEventCheckConfig struct {
 	CheckConfig *types.CheckConfig
 	Action      WatchActionType
-}
-
-// A WatchEventAsset contains the modified asset object and the action that occurred
-// during the modification.
-type WatchEventAsset struct {
-	Asset  *types.Asset
-	Action WatchActionType
 }
 
 // A WatchEventHookConfig contains the modified asset object and the action that occurred
@@ -217,12 +180,6 @@ type AssetStore interface {
 
 	// UpdateAsset creates or updates a given asset.
 	UpdateAsset(ctx context.Context, asset *types.Asset) error
-
-	// GetAssetWatcher returns a channel that emits WatchEventAsset structs notifying
-	// the caller that an Asset was updated. If the watcher runs into a terminal error
-	// or the context passed is cancelled, then the channel will be closed. The caller must
-	// restart the watcher, if needed.
-	GetAssetWatcher(ctx context.Context) <-chan WatchEventAsset
 }
 
 // AuthenticationStore provides methods for managing the JWT secret