+++ types/wrapper.go
@@ -7,6 +7,7 @@ import (
 	"path"
 	"reflect"
 	"strings"
+	"sync"
 
 	"github.com/sensu/sensu-go/api/core/v2"
 )
@@ -28,9 +29,12 @@ type rawWrapper struct {
 	Value      *json.RawMessage `json:"spec" yaml:"spec"`
 }
 
+// PackageMap contains a list of packages with their Resource Resolver func
 var packageMap = map[string]func(string) (Resource, error){
 	"core/v2": v2.ResolveResource,
 }
+
+var packageMapMu = &sync.RWMutex{}
 
 // UnmarshalJSON implements json.Unmarshaler
 func (w *Wrapper) UnmarshalJSON(b []byte) error {
@@ -42,6 +46,10 @@ func (w *Wrapper) UnmarshalJSON(b []byte) error {
 	if w.APIVersion == "" {
 		w.APIVersion = "core/v2"
 	}
+
+	// Guard read access to packageMap
+	packageMapMu.RLock()
+	defer packageMapMu.RUnlock()
 	resolver, ok := packageMap[w.TypeMeta.APIVersion]
 	if !ok {
 		return fmt.Errorf("invalid API version: %s", w.TypeMeta.APIVersion)
@@ -108,6 +116,13 @@ func WrapResource(r Resource) Wrapper {
 		ObjectMeta: r.GetObjectMeta(),
 		Value:      r,
 	}
+}
+
+// RegisterTypeResolver adds a package to packageMap with its resolver
+func RegisterTypeResolver(key string, resolver func(string) (Resource, error)) {
+	packageMapMu.Lock()
+	defer packageMapMu.Unlock()
+	packageMap[key] = resolver
 }
 
 func apiVersion(version string) string {