+++ CHANGELOG.md
@@ -13,13 +13,18 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 users to use a self-signed certificate without adding it to the operating
 system's CA store, either by explicitly trusting the signer, or by disabling
 TLS hostname verification.
+- Added a generic watcher in the store.
+- Added `RemoveProvider` method to authenticator.
 
 ### Changed
 - Removed unused workflow `rel_build_and_test` in CircleCI config.
+- Moved the `Provider` interface to `api/core/v2` package.
+- Moved the `Authenticator` interface to `backend/authentication` package.
 
 ### Fixed
 - Check TTL failure events are now much more reliable, and will persist even
 in the presence cluster member failures and cluster restarts.
+- Fix snakeCase version of keys in typeMap for acronyms.
 
 ## [5.1.1] - 2019-01-24